refactor: separate the specific logic about fieldsets in useCurrencyFieldset hook
