+++ src/components/currency-converter-form/TargetCurrencyFieldset/TargetCurrencyFieldset.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { useCurrentExchangeRate } from '@/lib/hooks';
+import { useCurrencyFieldset } from '@/lib/hooks';
 import {
   calculateRateToBaseCurrency,
   disableCurrencyFormField,
@@ -9,22 +9,17 @@ import {
 import { CurrencyConverterFormFieldsConfig } from '@/shared/types/globals';
 import styles from '@/styles/templates.module.scss';
 import { ChangeEventHandler, useCallback, useEffect } from 'react';
-import { useFormContext, useWatch } from 'react-hook-form';
+import { useFormContext } from 'react-hook-form';
 import { CurrencyError } from '../CurrencyError/CurrencyError';
 import { CurrencyInput } from '../CurrencyInput/CurrencyInput';
 import { CurrencySelect } from '../CurrencySelect/CurrencySelect';
 
 export const TargetCurrencyFieldset = (): React.JSX.Element => {
-  const { control, setValue, getValues } =
+  const { setValue, getValues } =
     useFormContext<CurrencyConverterFormFieldsConfig>();
 
-  const baseCurrency = useWatch({ control, name: 'baseCurrency' });
-  const targetCurrency = useWatch({ control, name: 'targetCurrency' });
-
-  const { data, isLoading, isFetching } = useCurrentExchangeRate({
-    baseCurrency,
-    targetCurrency,
-  });
+  const { targetCurrency, exchangeRateQueryResult } = useCurrencyFieldset();
+  const { data, isLoading, isFetching } = exchangeRateQueryResult;
 
   useEffect(() => {
     if (targetCurrency) {
