+++ src/lib/hooks/index.ts
@@ -1,1 +1,2 @@
+export { useCurrencyFieldset } from './useCurrencyFieldset.hook';
 export { useCurrentExchangeRate } from './useCurrentExchangeRate.hook';
